// apps/applications/saas-legacy/src/components/css-only/Fabrique/Button/types.ts
enum ButtonType {
  BUTTON = 'button',
  SUBMIT = 'submit',
  RESET = 'reset',
}

enum ButtonVariant {
  ICON = 'icon',
  OUTLINED = 'outlined',
  TEXT = 'text',
}

export enum ButtonSize {
  SMALL = 'small',
  MEDIUM = 'medium',
  LARGE = 'large',
}

export enum ButtonColor {
  PRIMARY = 'primary',
  SECONDARY = 'secondary',
}

export { ButtonType, ButtonVariant };